Corrección de validaciones y mejoras en vistas del sistema clínico

+++ database/migrations/2025_10_22_072718_create_nota_medicas_table.php
@@ -13,7 +13,7 @@ return new class extends Migration {
             $table->date('Fecha')->useCurrent();
             $table->time('Hora')->nullable();
             $table->float('Peso')->nullable();
-            $table->float('Talla')->nullable();
+            $table->decimal('Talla', 4, 2)->nullable(); 
             $table->string('Presion_Arterial', 20)->nullable();
             $table->integer('Frecuencia_Cardiaca')->nullable();
             $table->text('Exploracion_Fisica')->nullable();
